+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `FinishReason` enum
 - `Response` class `finish_reason` property
 - `Response` class `update_finish_reason` method
+### Changed
+- `Response` class typing bugs fixed
 ## [1.3] - 2026-07-28
 ### Added
 - `Session` class `enable_all_messages` method